requirements.txt
Django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.2.2
python-dotenv>=1.0.0
Pillow>=10.0.0